feat: sort HTTP headers
When printing the final output, PrintResult function iterate over the HTTP headers
stored in http.Header which uses map internally. This makes it difficult for a user
to parse the Headers in the output since keys in the map are not ordered.

This changes the behavior by using slice instead to store and sort the headers by name
and value. Additionally, removes header []value join logic from template.